+++ packages/components/src/ProgressCircular/IndeterminateProgress.tsx
@@ -28,7 +28,7 @@ import React, { FC } from 'react'
 import styled, { css, keyframes } from 'styled-components'
 import { progressCircularSVG, ProgressCircularSizes } from './size'
 import { progressCircularConstants } from './constants'
-import { CircleContainer } from './ProgessSvg'
+import { CircleContainer } from './ProgressSvg'
 
 interface IndeterminateSpinnerProps {
   size: ProgressCircularSizes
@@ -134,7 +134,6 @@ const leftSpin = keyframes`
     to {
       transform: rotate(265deg);
     }
-
 `
 
 const rightSpin = keyframes`
