+++ packages/components/src/index.ts
@@ -48,6 +48,7 @@ export * from './PageSize'
 export * from './Pagination'
 export * from './Popover'
 export * from './Portal'
+export * from './ProgressCircular'
 export * from './Spinner'
 export * from './Status'
 export * from './Table'
