+++ www/src/documentation/components/content/progress-circular.mdx
@@ -1,0 +1,10 @@
+---
+title: ProgressCircular
+storybook: true
+---
+
+```jsx
+<SpaceVertical gap="xsmall">
+  <ProgressCircular />
+</SpaceVertical>
+```
